+++ src/Types.hs
@@ -6,12 +6,10 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE InstanceSigs #-}
 {-# LANGUAGE KindSignatures #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE StrictData #-}
 {-# LANGUAGE BlockArguments #-}
-{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
@@ -21,7 +19,6 @@ module Types (
     Conditions (..),
     Config (..),
     ActionMode,
-    forConcurrentlyKi,
 ) where
 
 import Control.Exception.Safe (MonadCatch, MonadMask, MonadThrow)
@@ -33,16 +30,13 @@ import Control.Monad.Reader (
     asks,
     local,
  )
-import Control.Monad.Trans.Control (MonadBaseControl, StM, control)
-import Control.Monad.Base (MonadBase (liftBase))
+import Control.Monad.Trans.Control (MonadBaseControl)
 import Data.Set (Set)
 import Dhall (FromDhall (..))
 import GHC.Generics (Generic)
 import GHC.IO.Exception (ExitCode (..))
 import Katip (Katip (..), KatipContext (..), LogContexts, LogEnv (..), Namespace, Severity, Verbosity)
 import Path.Posix (Abs, Dir, Path)
-import Ki
-import Control.Concurrent.STM (atomically)
 
 
 newtype App a = MkApp
@@ -114,27 +108,3 @@ data Conditions = Conditions
     }
     deriving stock (Eq, Show, Generic, Ord)
     deriving anyclass (FromDhall)
-
-forConcurrentlyKi ::
-  (MonadBaseControl IO m, StM m (Ki.Thread b) ~ Ki.Thread b, StM m b ~ b, MonadIO m) =>
-  [a] ->
-  (a -> m b) ->
-  m [b]
-forConcurrentlyKi ns f = control $ \unlift -> scopedM \scope -> unlift $ do
-  threads <- mapM (forkM scope . f) ns
-  mapM (liftBase . atomically . Ki.await) threads
-
-forkM ::
-  (MonadBaseControl IO m, StM m (Ki.Thread a) ~ Ki.Thread a, StM m a ~ a) =>
-  Ki.Scope ->
-  m a ->
-  m (Ki.Thread a)
-forkM scope action =
-  control \unlift -> Ki.fork scope (unlift action)
-
-scopedM ::
-  (MonadBaseControl IO m,  StM m a ~ a) =>
-  (Ki.Scope -> m a) ->
-  m a
-scopedM action =
-  control \unlift -> Ki.scoped \scope -> unlift (action scope)